EIP-8024 dupn/swapn/exchange test fixes + small refactoring (dedicated util file)
